% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_btabletemplate.R
\name{use_btabletemplate}
\alias{use_btabletemplate}
\title{Open a basic template suitable for use with btable}
\usage{
use_btabletemplate(name, fmt = c("Rmd", "Rnw"), ..., open = TRUE)
}
\arguments{
\item{name}{name to save the file under}

\item{fmt}{character defining format (\code{Rmd} or \code{Rnw} are accepted)}

\item{...}{other options passed to \code{\link[usethis]{use_template}}}

\item{open}{logical, whether to open the file}
}
\value{
(by default) saves and opens a script of the chosen format
}
\description{
Open a basic template suitable for use with btable
}
\examples{
\dontrun{
# Rmd file
use_btabletemplate("foo")
# Rnw file
use_btabletemplate("foo", fmt = "Rnw")
}
}
